Reorganize GRPO/RL code into keys_values/rl module; add docs
Addresses review feedback on PR #128:
- Move GRPO files into a dedicated keys_values/rl/grpo package and the
  shared log-prob utility into keys_values/rl, so RL support is discoverable.
- Add docs/grpo_integration.md covering installation, how to run (standalone
  loop and TRL trainer), and where the KV cache is used (inference vs.
  gradient updates).
- Update all imports, tests, notebook, and AI-dev notes for the new paths.
