+++ ai_dev/trl_integration.md
@@ -5,8 +5,8 @@
 AI was used to generate docstrings for the functions and classes in the TRL
 integration code, prior to merging:
 
-- `keys_values/logprobs.py`
-- `keys_values/finetune/grpo.py`
-- `keys_values/finetune/grpo_loss.py`
-- `keys_values/finetune/grpo_loop.py`
-- `keys_values/generate/trl_rollout.py`
+- `keys_values/rl/logprobs.py`
+- `keys_values/rl/grpo/trainer.py`
+- `keys_values/rl/grpo/loss.py`
+- `keys_values/rl/grpo/loop.py`
+- `keys_values/rl/grpo/rollout.py`
